# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Angular Workshop API",
        "description": "API documentation for Angular Workshop application with categories, skills, exams, and user management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/exams": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "Get all exams",
                "operationId": "e063269a41b32018783ebc9989570c1b",
                "responses": {
                    "200": {
                        "description": "List of exams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exam"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Exams"
                ],
                "summary": "Create a new exam",
                "operationId": "0e7abe57c08bcf6118456fdae18d94f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_en",
                                    "name_ar",
                                    "desc_en",
                                    "desc_ar",
                                    "skill_id",
                                    "questions_no",
                                    "duration_mins",
                                    "difficulty"
                                ],
                                "properties": {
                                    "name_en": {
                                        "type": "string",
                                        "example": "Laravel Basics"
                                    },
                                    "name_ar": {
                                        "type": "string",
                                        "example": "أساسيات لارافيل"
                                    },
                                    "desc_en": {
                                        "type": "string",
                                        "example": "Test your Laravel knowledge"
                                    },
                                    "desc_ar": {
                                        "type": "string",
                                        "example": "اختبر معرفتك بلارافيل"
                                    },
                                    "skill_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "questions_no": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "duration_mins": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "example": "medium"
                                    },
                                    "img": {
                                        "type": "string",
                                        "example": "exam.png"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exam created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exams/{id}": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "Get a specific exam",
                "operationId": "6c844d345556bc42f5c7a112077a2880",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Exams"
                ],
                "summary": "Update an exam",
                "operationId": "f8244b2589f3c5669fb8b6f773a67593",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_en": {
                                        "type": "string",
                                        "example": "Laravel Basics"
                                    },
                                    "name_ar": {
                                        "type": "string",
                                        "example": "أساسيات لارافيل"
                                    },
                                    "desc_en": {
                                        "type": "string",
                                        "example": "Test your Laravel knowledge"
                                    },
                                    "desc_ar": {
                                        "type": "string",
                                        "example": "اختبر معرفتك بلارافيل"
                                    },
                                    "skill_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "questions_no": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "duration_mins": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "example": "medium"
                                    },
                                    "img": {
                                        "type": "string",
                                        "example": "exam.png"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Exam not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exams"
                ],
                "summary": "Delete an exam",
                "operationId": "fd9d48a3ffa9ff2aacdf338be34bcbff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exam deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exams/show-questions/{id}": {
            "get": {
                "tags": [
                    "Exams"
                ],
                "summary": "Get exam questions for taking",
                "operationId": "4f28a445ca327ad0216e070ccb2169eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam with questions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exams/start/{id}": {
            "post": {
                "tags": [
                    "Exams"
                ],
                "summary": "Start an exam",
                "operationId": "2fcf3c9c20a2cb931e6cd2ca6e8cf6bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "exam begins"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exams/submit/{id}": {
            "post": {
                "tags": [
                    "Exams"
                ],
                "summary": "Submit exam answers",
                "operationId": "8851bfdbe81505eab559abb739e5551e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answers"
                                ],
                                "properties": {
                                    "answers": {
                                        "type": "object",
                                        "example": {
                                            "1": 2,
                                            "2": 1,
                                            "3": 4
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "exam finished"
                                        },
                                        "score": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 85.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Exam not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}